% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{geom_label_pca}
\alias{geom_label_pca}
\title{Vector's Label for PCA bi-plot}
\usage{
geom_label_pca(
  pca_recipe,
  include_vars = NULL,
  x = PC1,
  y = PC2,
  geom_label_fun = ggplot2::geom_label,
  geom_label_args = list(color = "midnightblue"),
  ...
)
}
\arguments{
\item{pca_recipe}{Object class recipe that already \code{step_pca}}

\item{include_vars}{Variable to be included in the arrow, \code{NULL} (default) to include all variables.}

\item{x}{unquoted name to specify which principle component to plot in x-axis (must be one of PC1, PC2, PC3, ... ,etc.)}

\item{y}{unquoted name to specify which principle component to plot in y-axis (must be one of PC1, PC2, PC3, ... ,etc.)}

\item{geom_label_fun}{Labelling geom function}

\item{geom_label_args}{Args to \code{geom_label_fun}}

\item{...}{Args to \code{geom_label_fun}}
}
\value{
plot label
}
\description{
Vector's Label for PCA bi-plot
}
\examples{
library(lbmod)
prep_pca(iris) |>
  pca_scatter_plot(ggplot2::aes(color = Species)) +
  geom_label_pca(prep_pca(iris))
}
