% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{get_pc_percent_label}
\alias{get_pc_percent_label}
\title{Get vector of \%PC for label}
\usage{
get_pc_percent_label(pca_recipe, digits = 0)
}
\arguments{
\item{pca_recipe}{Object class recipe that already \code{step_pca}}

\item{digits}{Round digits}
}
\value{
A named vector
}
\description{
Intended to used in \code{ggplot2::labeller()}
}
