% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lbmod-package.R
\docType{package}
\name{lbmod-package}
\alias{lbmod}
\alias{lbmod-package}
\title{lbmod: Statistic & Modeling Function}
\description{
My wrapper package of various statistical and machine
    learning functions.
}
\author{
\strong{Maintainer}: Kittipos Sirivongrungson \email{ki11ip0.s.a.s@gmail.com}

}
\keyword{internal}
