% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{pca_load_plot}
\alias{pca_load_plot}
\title{Plot Loading Score from PCA}
\usage{
pca_load_plot(
  pca_recipe,
  num_comp = 1:4,
  num_top_vars = Inf,
  fill = c("#b6dfe2", "#0A537D"),
  nrow = NULL,
  ncol = NULL,
  ...
)
}
\arguments{
\item{pca_recipe}{Object class recipe that already \code{step_pca}}

\item{num_comp}{Integer vector indicate which PCs to show.}

\item{num_top_vars}{Integer indicate number of top variables, ranking by absolute loading score, to show. Default \code{Inf} means all variables.}

\item{fill}{(character) Fill of positive & negative values}

\item{nrow}{(numeric) Number of rows in facet}

\item{ncol}{(numeric) Number of columns in facet}

\item{...}{passed to \code{geom_col}}
}
\value{
plot
}
\description{
Plot Loading Score from PCA
}
\examples{
library(magrittr)
library(lbmod)
iris \%>\%
  prep_pca() \%>\%
  pca_load_plot()
}
