% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{pca_load_plot_all}
\alias{pca_load_plot_all}
\title{Plot All Loading Score from PCA}
\usage{
pca_load_plot_all(pca_recipe, num_comp = 1:4)
}
\arguments{
\item{pca_recipe}{Object class recipe that already \code{step_pca}}

\item{num_comp}{Integer vector indicate which PCs to show.}
}
\value{
Plot
}
\description{
Plot all of the variables (loading) in a single plot.
}
\examples{
library(magrittr)
library(lbmod)
iris \%>\%
  prep_pca() \%>\%
  pca_load_plot_all()
}
