% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univar_multi.R
\name{univar_multi_bare}
\alias{univar_multi_bare}
\title{Multiple Pairs of Univariate Analysis (No Processing)}
\usage{
univar_multi_bare(data, fun = "lm", y_var, x_vars, pkg = "stats", ...)
}
\arguments{
\item{data}{A data frame}

\item{fun}{(character) A modeling function}

\item{y_var}{(character) Outcome variable}

\item{x_vars}{(character) One or more predictor variable(s)}

\item{pkg}{(character) Package of \code{fun}}

\item{...}{Passing to \code{fun}}
}
\value{
A list with names as \code{x_vars}
}
\description{
Fitting multiple univariate models of one \code{y_var} on one or more \code{x_vars} using any \code{fun}.
The output is bare (no modified)
}
\examples{
iris |> univar_multi_bare(fun = "lm", "Sepal.Length", "Species")

}
