% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univar_multi.R
\name{univar_multi_tidy}
\alias{univar_multi_tidy}
\title{Multiple Pairs of Univariate Analysis with Tidied Output}
\usage{
univar_multi_tidy(data, fun = "lm", y_var, x_vars, pkg = "stats", ...)
}
\arguments{
\item{data}{A data frame}

\item{fun}{(character) A modeling function}

\item{y_var}{(character) Outcome variable}

\item{x_vars}{(character) One or more predictor variable(s)}

\item{pkg}{(character) Package of \code{fun}}

\item{...}{Passing to \code{fun}}
}
\value{
A \code{broom::tidy} data frame
}
\description{
Fitting multiple univariate models of one \code{y_var} on one or more \code{x_vars} using any \code{fun}.
The output will be \code{broom::tidy}.
}
\examples{
iris |> univar_multi_tidy(fun = "lm", "Sepal.Length", c("Species", "Petal.Width"))

}
